Adding all under the agent script and removing nodes for cleanliness
